+++ lib/matrix_sdk/util/extensions.rb
@@ -73,8 +73,9 @@ module MatrixSdk
   module Logging
     def logger
       return MatrixSdk.logger if MatrixSdk.global_logger?
+      return @logger if instance_variable_defined?(:@logger) && @logger
 
-      @logger || ::Logging.logger[self]
+      ::Logging.logger[self]
     end
 
     def logger=(logger)